\name{subselect-internal}
\title{Internal subselect Objects}

\alias{validation}
\alias{validgenetic}
\alias{validannimp}

\description{
  Internal subselect functions that validate input to the search functions.
}
\usage{

validation(mat, kmin, kmax, exclude, include, criterion, pcindices, tolval)
validgenetic(kmin, kmax, popsize, mutprob, exclude, nexclude, include, ninclude, initialpop)
validannimp(kmin, kmax, nsol, exclude, nexclude, include, ninclude, initialsol,implog)
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
